moved Entity classes to entity package
