+++ data/db/src/main/java/com/freshdigitable/udonroad2/data/db/entity/MemberListListEntity.kt
@@ -1,0 +1,45 @@
+/*
+ * Copyright (c) 2020. Matsuda, Akihit (akihito104)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.freshdigitable.udonroad2.data.db.entity
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.ForeignKey
+import androidx.room.PrimaryKey
+import com.freshdigitable.udonroad2.model.MemberListId
+
+@Entity(
+    tableName = "member_list_list",
+    foreignKeys = [
+        ForeignKey(
+            entity = MemberListEntity::class,
+            parentColumns = ["id"],
+            childColumns = ["member_list_id"]
+        )
+    ]
+)
+internal data class MemberListListEntity(
+    @ColumnInfo(name = "member_list_id", index = true)
+    val memberListId: MemberListId,
+
+    @PrimaryKey(autoGenerate = true)
+    @ColumnInfo(name = "order")
+    val order: Long = 0,
+
+    @ColumnInfo(name = "owner")
+    val owner: String
+)
